Refactor: Rename enem_ai_analyst to ai_data_analyst
Renamed the main application directory and all references from `enem_ai_analyst` to `ai_data_analyst` to reflect the project's more generic capabilities.

Changes include:
- Renamed `enem_ai_analyst/` directory to `ai_data_analyst/`.
- Updated Python import statements and module path references in all `.py` files.
- Changed the package `name` in `pyproject.toml`.
- Updated paths and names in `README.md`.
- Updated `app_name` in JSON test fixtures.
- Performed a basic find/replace for the project name in `.md` documentation files in `docs/` (filenames in `docs/` and full content overhaul are pending future work).

Application launches successfully after these changes. Project tests are timing out, which may indicate a separate issue with the test suite or environment.
