+++ tests/integration/test_analysis_agent.py
@@ -5,6 +5,6 @@ from google.adk.evaluation.agent_evaluator import AgentEvaluator
 async def test_analysis_agent_evals():
     """Test the analysis agent using the defined evalset."""
     await AgentEvaluator.evaluate(
-        agent_module="enem_ai_analyst.sub_agents.analysis_agent",
+        agent_module="ai_data_analyst.sub_agents.analysis_agent",
         eval_dataset_file_path_or_dir="tests/integration/fixtures/analysis_agent/eval_set_analysis_agent.test.json"
     )
