+++ tests/integration/test_narrative_agent.py
@@ -5,6 +5,6 @@ from google.adk.evaluation.agent_evaluator import AgentEvaluator
 async def test_narrative_agent_evals():
     """Test the narrative agent using the defined evalset."""
     await AgentEvaluator.evaluate(
-        agent_module="enem_ai_analyst.sub_agents.narrative_agent",
+        agent_module="ai_data_analyst.sub_agents.narrative_agent",
         eval_dataset_file_path_or_dir="tests/integration/fixtures/narrative_agent/eval_set_narrative_agent.test.json"
     )
